Add benchmarking for has table performance
To see the difference in speed between open addressing and chaining implementations.

Open addressing performs well at low load factors because there is a high chance that the probed slot is free but massively degrades at the higher load factors due to more probing.

Chaining is more consistent and degrades more gradually, so it is more suitable for the higher load factors.

See `benchmark_results.md` for output.